Make the project cross-build across scala 2 and 3
