+++ encoder-akka/src/main/scala-2/com/dispalt/taglessAkka/akkaEncoder.scala
@@ -2,7 +2,7 @@ package com.dispalt.taglessAkka
 
 import com.dispalt.tagless.EncoderGeneratorMacro
 
-import scala.annotation.{compileTimeOnly, StaticAnnotation}
+import scala.annotation.{StaticAnnotation, compileTimeOnly}
 import scala.reflect.macros.whitebox
 
 @compileTimeOnly("Cannot expand @akkaEncoder")
