+++ encoder-akka/src/main/scala-3/com/dispalt/taglessAkka/AnyValGenerator.scala
@@ -1,0 +1,3 @@
+package com.dispalt.taglessAkka
+
+trait AnyValGenerator extends DefaultGenerator {}
