+++ encoder-akka/src/main/scala/com/dispalt/taglessAkka/AkkaCodecFactory.scala
@@ -3,7 +3,7 @@ package com.dispalt.taglessAkka
 import com.dispalt.tagless.util.{CodecFactory, WireProtocol}
 
 object AkkaCodecFactory extends CodecFactory[AkkaImpl] {
-  override def encode[A](implicit p: AkkaImpl[A]): WireProtocol.Encoder[A] = { a: A =>
+  override def encode[A](implicit p: AkkaImpl[A]): WireProtocol.Encoder[A] = { (a: A) =>
     p.encode(a)
   }
