+++ encoder-akka/src/main/scala/com/dispalt/taglessAkka/AkkaImpl.scala
@@ -1,14 +1,12 @@
 package com.dispalt.taglessAkka
 
-import java.nio.ByteBuffer
-
 import akka.actor.ActorSystem
 import akka.serialization.SerializationExtension
 
+import java.nio.ByteBuffer
 import scala.annotation.implicitNotFound
 import scala.reflect.ClassTag
 import scala.util.Try
-import scala.reflect.macros.blackbox.{Context => MacroContext}
 
 @implicitNotFound(
   "Could not find implicit AkkaImpl[${A}] Which usually means you do not have an implicit akka.actor.ActorSystem in scope."
@@ -55,10 +53,6 @@ object AkkaImpl extends AnyValGenerator {
 
 }
 
-trait AnyValGenerator extends DefaultGenerator {
-  implicit def akkaImplMacroAnyVal[A <: AnyVal]: AkkaImpl[A] = macro AnyValGeneratorMacros.impl[A]
-}
-
 trait DefaultGenerator {
 
   implicit def akkaImplGen[A <: AnyRef](implicit system: ActorSystem, ct: ClassTag[A]): AkkaImpl[A] = {
@@ -73,28 +67,3 @@ trait DefaultGenerator {
   }
 }
 
-class AnyValGeneratorMacros(val c: MacroContext) {
-  import c.universe._
-
-  private def withAnyValParam[R](tpe: Type)(f: Symbol => R): Option[R] =
-    tpe.baseType(c.symbolOf[AnyVal]) match {
-      case NoType => None
-      case _ =>
-        primaryConstructor(tpe).map(_.paramLists.flatten).collect {
-          case param :: Nil => f(param)
-        }
-    }
-
-  private def primaryConstructor(t: Type) =
-    t.members.collectFirst {
-      case m: MethodSymbol if m.isPrimaryConstructor => m.typeSignature.asSeenFrom(t, t.typeSymbol)
-    }
-
-  def impl[T <: AnyVal](implicit t: WeakTypeTag[T]): Expr[AkkaImpl[T]] = {
-    c.Expr[AkkaImpl[T]](withAnyValParam(t.tpe) { param =>
-      q"""
-        implicitly[_root_.com.dispalt.taglessAkka.AkkaImpl[${param.typeSignature}]].imap(new ${t.tpe}(_))((v: ${t.tpe}) => v.${param.name.toTermName})
-      """
-    }.getOrElse(c.abort(c.enclosingPosition, s"Could find ${t.tpe}")))
-  }
-}
