+++ encoder-boopickle/src/main/scala-2/com/dispalt/taglessBoopickle/BoopickleWireProtocol.scala
