+++ encoder-kryo/src/test/scala-2/com/dispalt/taglessKryo/tests/SafeAlg.scala
