+++ encoder-macros/src/main/scala-3/com/dispalt/tagless/EncoderGeneratorMacro.scala
@@ -1,0 +1,1 @@
+package com.dispalt.tagless
