+++ encoder-macros/src/test/scala-2/com/dispalt/tagless/TwoWaySimulator.scala
