+++ macros/src/test/scala-2/cats/tagless/tests/SafeAlg.scala
