+++ tests/src/test/scala-2/cats/tagless/tests/CatsTaglessTestSuite.scala
